+++ config/admin-bar.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Admin Bar Runtime Configuration Parameters.
+ *
+ * @package     KnowTheCode\UpDevTools\Admin
+ * @since       1.0.3
+ * @author      hellofromTonya
+ * @link        https://KnowTheCode.io
+ * @license     GNU-2.0+
+ */
+namespace KnowTheCode\UpDevTools\Admin;
+
+return array(
+	'message' => 'UPDEVTOOLS ACTIVE',
+	'colors'  => array(
+		'admin_bar_background_color' => '#627f00',
+		'message_background_color'   => '#cb4b14',
+		'message_hover_color'        => '#1b202d',
+	),
+);